CreateUserController.java
Changed name on method from createUserAction to createUser.

Changed the order on the UserValidation fields in the method to encapsulate better.

Added some more logic, the method creates a user-object and an empty noteOverview object for the new user.

setScene-method will be called when pressing create user button, and the user will be sent back to login-scene.

Also added the methos loadCreateUserInfo, but have not implemented any logic here yet.

CreateUserController class extends AbstractController now.

Co-authored-by: Åsmund Løvoll <aasmul@stud.ntnu.no>